requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.6.0
blis==0.7.8
boto3==1.26.115
botocore==1.29.115
bs4==0.0.1
catalogue==2.0.8
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
cymem==2.0.6
datasets==2.4.0
dill==0.3.5.1
emoji==2.0.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.36.0
frozenlist==1.3.1
fsspec==2022.7.1
graphviz==0.20.1
hdt==2.3
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
lmdb==1.3.0
marisa-trie==0.7.7
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
mwparserfromhell==0.6.4
mypy-extensions==0.4.3
networkx==2.8.5
nltk==3.7
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pathy==0.6.2
patsy==0.5.3
Pillow==9.2.0
platformdirs==2.5.2
preshed==3.0.7
protobuf==3.20.1
pyarrow==9.0.0
pybind11==2.2.4
pycountry==22.3.5
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-transformers==1.2.0
pytz==2022.2.1
PyYAML==6.0
rank-bm25==0.2.2
regex==2022.8.17
requests==2.28.1
responses==0.18.0
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.0
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stanza==1.4.0
statsmodels==0.13.5
tensorboardX==2.5.1
thinc==8.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
torch-transformers==1.0.4
torchviz==0.0.2
tqdm==4.64.0
transformers==4.21.1
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
Werkzeug==2.2.2
wikimapper==0.1.6
wikipedia2vec==1.0.5
wikitables==0.5.5
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1